DASH Reporting ERROR when seeking to end of file
SIDX source was not created when seeking to end of file
so we had to handle it properly when returning null SIDX.
Also when seeking to the end we passed all periods which
resulted in using the first period instead of the last one.
When seeking to end we did not create a fetcher for audio
which resulted in AudioThread never getting the EOS AccessUnit.
Created a flag that if we seek to the end, videofetcher will
report EOS and set an internal flag that we are EOS so future
calls to dequeueAccessUnit will return AccessUnit.EOS.