Add DisplayManager to OutputController
Adds a DisplayManager and a DisplayListener to the OutputController
and changes the responsibility of managing displays to
the DisplayManager instead of having the MediaRouter to handle this.
Fixes an issue where if displays with restrictions were added during
playback we could not detect this and send the appropriate callback
to onOutputBlocked().