Update Api tests
Update the Api tests to be able to run them for multiple
content based on certain conditions, annotations is used
to describe what type of conditions and content a test
need to be able to run.
To be able to run with multiple content the new runner
MultipleContentTestRunner must be used to start the test.
If no content is supplied to the Api test it will fall
back and use a default content, just like before.