Make HandlerHelper non static
HandlerHelper was static and used from several classes
and player instances. There is a risk that calls get stuck
if release is called on the player at same time as blocking
function using sendMessageAndWait is called.
Add possibility to quit and return from sendMessageAndAwaitResponse.
HandlerHelper then need to be non static to not affect
multiple player instances.