MPEG-DASH: Delay switch down
If the bandwidth drops below the required bandwidth a switch to
a lower representation is done immediately. Sometimes only one
reading is below the required bandwidth and the next reading is
above again. Since the switch up requires more than 100% we might not
switch back up again. Avoid this by allowing up to 3 consecutie
readings to be between 90% and 100% without switching down.