DASH: Check for duration to determine number of segments
Previously for DASH the number of segments were determined
by dividing the media length by the length of a segment.
This could cause error when playing files where the last
segment were larger then previous segments since we expected
one segment more then there actually existed. Now we instead
compare the time for an access unit and if it is larger than
the duration or, if the time for the next fragment is larger
than the duration we know that the last segment is playing. This
is only used when downloading mpd files with a segment template
without a segment timeline.