OutputControl: Delay sending MSG_UPDATE after display added
Delays the message to the event handler by 1000ms to ensure
that hdcp flag is set. Without the delay this could cause an
issue where OnOutputBlocked events were passed on since the
flag for hdcp had not yet been set even though the content
was permitted to play on an external HDMI display.