Reconnect handling in Paused (non-consuming) state
If MediaPlayer is in paused state and device is moved out of range
we will now not send SOURCE_ERROR. While we are not consuming
anything from buffer we will try to reconnect. If play is called
we will play out our buffer and try to reconnect while we still
have data left in buffer. If we run out of data we will send
SOURCE_ERROR.