Update file browsing in Demo application
File browsing has been updated to be more intuitive. Directories
are listed before files and now end with '/' so they are
easier to distinguish from files. Filenames that are too long
to fit on one line are now wrapped to show on multiple lines
as is done in the media store browsing view.